+++ app/src/main/java/com/xinhui/mobileprojectweatherapp/data/model/ForeCastData.kt
@@ -36,7 +36,7 @@ data class ForecastData(
 ){
     fun toForecastWeatherDisplay() :ForecastWeatherDisplay{
         return ForecastWeatherDisplay(
-            temp,
+            temp.toInt(),
             weather,
             timestamp_local
         )
